requirements.txt
mcp>=1.0.0
httpx>=0.27.0
openai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
